fixed print time
