+++ exchange/exchange.py
@@ -9,6 +9,7 @@ from collections import deque
 
 from OuchServer.ouch_messages import OuchClientMessages, OuchServerMessages
 from OuchServer.ouch_server import nanoseconds_since_midnight
+from OuchServer.ouch_server import printTime 
 from exchange.order_store import OrderStore
 
 ###
@@ -49,15 +50,6 @@ class Exchange:
             OuchClientMessages.CancelOrder: self.cancel_order_atomic,
             OuchClientMessages.SystemEvent: self.system_event_atomic,  #jason
             OuchClientMessages.ModifyOrder: None}
-    
-    def printTime(nanoseconds):#jason
-        out = ""
-        millis  = Math.floor((nanoseconds / 1000000) % 1000)
-        seconds = Math.floor((nanoseconds / 1000000000) % 60)
-        minutes = Math.floor(nanoseconds / (60*1000000000) % 60)
-        hours   = Math.floor(nanoseconds / (60*60*1000000000) % 24)
-        out = "[" + hours + ":" + minutes + ":" + seconds + ":" + millis + "]"
-        return out
 
     def system_event_atomic(self, system_event_message, timestamp):      #jason
         if system_event_message['event_code'] == b'E':
